Support for g54 offset.  Pressing Shift-Home will set the g54 offset for the active axis to the current position (the relative coordinate goes to zero.)
This could be enhanced to allow specification of the new relative
coordinate. For instance, touch the .002 feeler gauge on top of the
work, then set the g54 offset for Z to the current Z + .002.

Also, when displaying relative coordinates, move the axes to the
work origin, and because I have no artistic skills, put a stupid
snowflake thing at the machine origin.